half way

+++ app/Controller/TopicsController.php
@@ -31,6 +31,10 @@ class TopicsController extends AppController {
 	}
 
 
+	public function admin_index() {
+		$this->Topic->recursive = 0;
+		$this->set('topics', $this->Paginator->paginate());
+	}
 
 /**
  * view method
@@ -47,15 +51,24 @@ class TopicsController extends AppController {
 		$this->set('topic', $this->Topic->find('first', $options));
 	}
 
+	public function admin_view($id = null) {
+		if (!$this->Topic->exists($id)) {
+			throw new NotFoundException(__('Invalid topic'));
+		}
+		$options = array('conditions' => array('Topic.' . $this->Topic->primaryKey => $id));
+		$this->set('topic', $this->Topic->find('first', $options));
+	}
+
 /**
  * add method
  *
  * @return void
  */
-	public function add() {
+	public function admin_add() {
+		if($this->isAuthorized()){
 		if ($this->request->is('post')) {
 			$this->Topic->create();
-			$this->request->data['Topic']['user_id'] = Authcomponent::user('role');
+			$this->request->data['Topic']['user_id'] = Authcomponent::user('id');
 			if ($this->Topic->save($this->request->data)) {
 				$this->Session->setFlash(__('The topic has been saved.'));
 				return $this->redirect(array('action' => 'index'));
@@ -63,6 +76,10 @@ class TopicsController extends AppController {
 				$this->Session->setFlash(__('The topic could not be saved. Please, try again.'));
 			}
 		}
+		} else{
+			$this->Session->setFlash(__('You do not have permission to do this'));
+			return $this->redirect(array('controller'=>'users','action' => 'admin_dashboard'));
+			}
 		$users = $this->Topic->User->find('list');
 		$this->set(compact('users'));
 	}
@@ -74,21 +91,26 @@ class TopicsController extends AppController {
  * @param string $id
  * @return void
  */
-	public function edit($id = null) {
-		if (!$this->Topic->exists($id)) {
-			throw new NotFoundException(__('Invalid topic'));
-		}
-		if ($this->request->is(array('post', 'put'))) {
-			if ($this->Topic->save($this->request->data)) {
-				$this->Session->setFlash(__('The topic has been saved.'));
-				return $this->redirect(array('action' => 'index'));
+	public function admin_edit($id = null) {
+		if($this->isAuthorized()){
+			if (!$this->Topic->exists($id)) {
+				throw new NotFoundException(__('Invalid topic'));
+				}
+			if ($this->request->is(array('post', 'put'))) {
+				if ($this->Topic->save($this->request->data)) {
+					$this->Session->setFlash(__('The topic has been saved.'));
+					return $this->redirect(array('action' => 'index'));
+				} else {
+					$this->Session->setFlash(__('The topic could not be saved. Please, try again.'));
+						}
 			} else {
-				$this->Session->setFlash(__('The topic could not be saved. Please, try again.'));
+				$options = array('conditions' => array('Topic.' . $this->Topic->primaryKey => $id));
+				$this->request->data = $this->Topic->find('first', $options);}
+		} else{
+			$this->Session->setFlash(__('You do not have permission to do this'));
+			return $this->redirect(array('controller'=>'users','action' => 'admin_dashboard'));
 			}
-		} else {
-			$options = array('conditions' => array('Topic.' . $this->Topic->primaryKey => $id));
-			$this->request->data = $this->Topic->find('first', $options);
-		}
+		
 		$users = $this->Topic->User->find('list');
 		$this->set(compact('users'));
 	}
@@ -100,17 +122,23 @@ class TopicsController extends AppController {
  * @param string $id
  * @return void
  */
-	public function delete($id = null) {
-		$this->Topic->id = $id;
-		if (!$this->Topic->exists()) {
-			throw new NotFoundException(__('Invalid topic'));
+	public function admin_delete($id = null) {
+		if($this->isAuthorized()){
+			$this->Topic->id = $id;
+			if (!$this->Topic->exists()) {
+				throw new NotFoundException(__('Invalid topic'));
+			}
+	
+			if ($this->Topic->delete()) {
+				$this->Session->setFlash(__('The topic has been deleted.'));
+			} else {
+				$this->Session->setFlash(__('The topic could not be deleted. Please, try again.'));
+					}
+			return $this->redirect(array('action' => 'index'));
+		}else{
+			$this->Session->setFlash(__('You do not have permission to do this'));
+			return $this->redirect(array('controller'=>'users','action' => 'admin_dashboard'));
+			}
 		}
-		$this->request->allowMethod('post', 'delete');
-		if ($this->Topic->delete()) {
-			$this->Session->setFlash(__('The topic has been deleted.'));
-		} else {
-			$this->Session->setFlash(__('The topic could not be deleted. Please, try again.'));
-		}
-		return $this->redirect(array('action' => 'index'));
-	}
+
 }
